+++ setup.py
@@ -15,7 +15,7 @@ setup(name='tinygrad',
       long_description=long_description,
       long_description_content_type='text/markdown',
       packages = ['tinygrad', 'tinygrad.runtime.autogen', 'tinygrad.codegen', 'tinygrad.nn', 'tinygrad.renderer', 'tinygrad.engine',
-                  'tinygrad.runtime', 'tinygrad.runtime.driver', 'tinygrad.runtime.graph', 'tinygrad.shape', 'tinygrad.features'],
+                  'tinygrad.runtime', 'tinygrad.runtime.driver', 'tinygrad.runtime.graph', 'tinygrad.shape'],
       classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: MIT License"
@@ -59,7 +59,8 @@ setup(name='tinygrad',
             "mkdocs-material",
             "mkdocstrings[python]",
             "markdown-callouts",
-            "markdown-exec[ansi]"
+            "markdown-exec[ansi]",
+            "black"
         ],
         'testing_tf': [
             "tensorflow==2.15.1",
